src/main/java/com/controller/GameController.java
package com.controller;

import com.service.GameService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RestController;

@RestController
public class GameController {
    private GameService gameService;

    @Autowired
    public GameController(GameService gameService) {
        this.gameService = gameService;
    }

    @RequestMapping(value = "/game/nextGeneration", method = RequestMethod.POST)
    public  String getNextGeneration(@RequestBody String initialState){
        return gameService.getNextGeneration(initialState);
    }
}
